refactor: 예약 query 서비스 분리
