+++ src/test/java/roomescape/service/ReservationServiceTest.java
@@ -13,7 +13,6 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.boot.test.context.SpringBootTest.WebEnvironment;
 import org.springframework.test.context.jdbc.Sql;
-import roomescape.dto.reservation.ReservationFilter;
 import roomescape.dto.reservation.ReservationRequest;
 import roomescape.dto.reservation.ReservationResponse;
 import roomescape.dto.reservationtime.ReservationTimeResponse;
@@ -26,6 +25,9 @@ class ReservationServiceTest {
 
     @Autowired
     ReservationService reservationService;
+
+    @Autowired
+    ReservationQueryService reservationQueryService;
 
     @Autowired
     ReservationTimeService reservationTimeService;
@@ -90,20 +92,9 @@ class ReservationServiceTest {
     }
 
     @Test
-    void 존재하지_않는_id로_조회할_경우_예외_발생() {
-        //given
-        List<ReservationResponse> allReservations = reservationService.getAllReservations();
-        Long notExistIdToFind = allReservations.size() + 1L;
-
-        //when, then
-        assertThatThrownBy(() -> reservationService.getReservation(notExistIdToFind))
-                .isInstanceOf(IllegalArgumentException.class);
-    }
-
-    @Test
     void 존재하지_않는_id로_삭제할_경우_예외_발생() {
         //given
-        List<ReservationResponse> allReservations = reservationService.getAllReservations();
+        List<ReservationResponse> allReservations = reservationQueryService.getAllReservations();
         Long notExistIdToFind = allReservations.size() + 1L;
 
         //when, then
@@ -114,14 +105,14 @@ class ReservationServiceTest {
     @Test
     void 예약_삭제_시_예약_대기가_존재한다면_우선순위가_제일_높은_예약_대기를_자동_승인() {
         // given
-        List<ReservationResponse> beforeReservations = reservationService.getAllReservations();
+        List<ReservationResponse> beforeReservations = reservationQueryService.getAllReservations();
         ReservationResponse beforeFirstReservation = beforeReservations.get(0);
 
         // when
         reservationService.deleteReservation(beforeFirstReservation.id());
 
         // then
-        List<ReservationResponse> afterReservations = reservationService.getAllReservations();
+        List<ReservationResponse> afterReservations = reservationQueryService.getAllReservations();
         ReservationResponse afterFirstReservation = afterReservations.get(0);
         assertThat(beforeFirstReservation).isNotEqualTo(afterFirstReservation);
     }
@@ -129,7 +120,7 @@ class ReservationServiceTest {
     @Test
     void 예약_삭제_시_예약_대기가_존재하지_않으면_예약_정상_삭제() {
         // given
-        List<ReservationResponse> reservations = reservationService.getAllReservations();
+        List<ReservationResponse> reservations = reservationQueryService.getAllReservations();
         ReservationResponse firstReservations = reservations.get(0);
         reservationService.deleteReservation(firstReservations.id());
 
@@ -137,50 +128,7 @@ class ReservationServiceTest {
         reservationService.deleteReservation(firstReservations.id());
 
         // then
-        List<ReservationResponse> afterReservations = reservationService.getAllReservations();
+        List<ReservationResponse> afterReservations = reservationQueryService.getAllReservations();
         assertThat(afterReservations).isEmpty();
-    }
-
-    @Sql("/reservation-filter-api-test-data.sql")
-    @Test
-    void 특정_사용자로_필터링_후_예약_조회() {
-        //given
-        Long filteringUserId = 1L;
-        ReservationFilter reservationFilter = new ReservationFilter();
-        reservationFilter.setMemberId(filteringUserId);
-
-        //when
-        List<ReservationResponse> reservationResponses = reservationService.getReservationsByFilter(reservationFilter);
-
-        //then
-        boolean isAllMatch = reservationResponses.stream()
-                .allMatch(response -> response.member().id() == filteringUserId);
-        assertThat(isAllMatch).isTrue();
-    }
-
-    @Sql("/reservation-filter-api-test-data.sql")
-    @Test
-    void 특정_테마와_날짜로_필터링_후_예약_조회() {
-        //given
-        Long filteringThemeId = 1L;
-        LocalDate dateFrom = LocalDate.of(2024, 5, 2);
-        LocalDate dateTo = LocalDate.of(2024, 5, 3);
-
-        ReservationFilter reservationFilter = new ReservationFilter();
-        reservationFilter.setThemeId(filteringThemeId);
-        reservationFilter.setDateFrom(dateFrom);
-        reservationFilter.setDateTo(dateTo);
-
-        //when
-        List<ReservationResponse> reservationResponses = reservationService.getReservationsByFilter(reservationFilter);
-
-        //then
-        boolean isAllMatch = reservationResponses.stream()
-                .allMatch(response ->
-                        response.theme().id() == filteringThemeId &&
-                        (response.date().isEqual(dateFrom) || response.date().isAfter(dateFrom)) &&
-                        (response.date().isEqual(dateTo) || response.date().isBefore(dateTo))
-                );
-        assertThat(isAllMatch).isTrue();
     }
 }
